refactor (fe): Refactor 'courses' routes and components to 'lms'
Renamed all frontend routes, pages, and navigation from '/courses' and related variables to '/lms' for improved clarity and consistency. Updated navigation links, route definitions, and all internal references in components and pages to use the new 'lms' naming convention. This change unifies the terminology across the application and prepares for broader LMS functionality.
