--- frontend/src/routes/(protected)/courses/$courseCode/assignments/index.tsx
@@ -1,12 +1,0 @@
-import AssignmentPage from '@/pages/shared/courses/$courseCode/assignments'
-import { createFileRoute } from '@tanstack/react-router'
-
-export const Route = createFileRoute(
-  '/(protected)/courses/$courseCode/assignments/',
-)({
-  component: RouteComponent,
-})
-
-function RouteComponent() {
-  return <AssignmentPage />
-}
