--- frontend/src/routes/(protected)/courses/$courseCode/modules/edit.tsx
@@ -1,23 +1,0 @@
-import RoleComponentManager from '@/components/role-component-manager'
-import { useAuth } from '@/features/auth/auth.hook'
-import ModulesCreateAdminPage from '@/pages/admin/courses/$courseCode/modules/create'
-import { createFileRoute } from '@tanstack/react-router'
-
-export const Route = createFileRoute(
-  '/(protected)/courses/$courseCode/modules/edit',
-)({
-  component: RouteComponent,
-})
-
-function RouteComponent() {
-  const { authUser } = useAuth('protected')
-
-  return (
-    <RoleComponentManager
-      currentRole={authUser.role}
-      roleRender={{
-        admin: <ModulesCreateAdminPage />,
-      }}
-    />
-  )
-}
