--- frontend/src/routes/(protected)/courses/$courseCode_.publish.tsx
@@ -1,39 +1,0 @@
-import { createFileRoute, useParams } from '@tanstack/react-router'
-import { useAuth } from '@/features/auth/auth.hook.ts'
-import RoleComponentManager from '@/components/role-component-manager.tsx'
-import CoursePublisher from '@/pages/admin/courses/$courseCode/edit/course-publisher.tsx'
-
-export const Route = createFileRoute(
-  '/(protected)/courses/$courseCode_/publish',
-)({
-  validateSearch: (search: Record<string, unknown>) => {
-    return {
-      scheduled: search.scheduled === 'true',
-      publish: search.publish === 'true',
-      unpublish: search.unpublish === 'true',
-    }
-  },
-  component: RouteComponent,
-})
-
-function RouteComponent() {
-  const { authUser } = useAuth('protected')
-  const { courseCode } = useParams({ strict: false })
-  const { scheduled, publish, unpublish } = Route.useSearch()
-
-  if (!authUser || !courseCode) {
-    //TODO: or a loader / skeleton
-    return null
-  }
-
-  return (
-    <RoleComponentManager
-      currentRole={authUser.role}
-      roleRender={{
-        admin: (
-          <CoursePublisher courseCode={courseCode} scheduled={scheduled} />
-        ),
-      }}
-    />
-  )
-}
