+++ frontend/src/routes/(protected)/lms/$lmsCode/modules/$itemId/index.tsx
@@ -1,0 +1,16 @@
+import ModulesContentPage from '@/pages/shared/lms/$lmsCode/modules/$itemId'
+import { createFileRoute } from '@tanstack/react-router'
+
+export const Route = createFileRoute(
+  '/(protected)/lms/$lmsCode/modules/$itemId/',
+)({
+  component: RouteComponent,
+})
+
+//TODO: Replace with actual fetch
+function RouteComponent() {
+  
+  return (
+    <ModulesContentPage />
+  )
+}
