+++ frontend/src/routes/(protected)/lms/$lmsCode/modules/$itemId/submit.tsx
@@ -1,0 +1,13 @@
+import SubmissionPage from '@/pages/admin/lms/$lmsCode/modules/$itemId/submit'
+import { createFileRoute } from '@tanstack/react-router'
+
+export const Route = createFileRoute(
+  '/(protected)/lms/$lmsCode/modules/$itemId/submit',
+)({
+  component: RouteComponent,
+})
+
+function RouteComponent() {
+  const { itemId } = Route.useParams()
+  return <SubmissionPage assignmentId={itemId} />
+}
