+++ frontend/src/routes/(protected)/lms/$lmsCode/modules/create.tsx
@@ -1,0 +1,23 @@
+import RoleComponentManager from '@/components/role-component-manager.tsx'
+import { useAuth } from '@/features/auth/auth.hook.ts'
+import ModulesCreateAdminPage from '@/pages/admin/lms/$lmsCode/modules/create'
+import { createFileRoute } from '@tanstack/react-router'
+
+export const Route = createFileRoute(
+  '/(protected)/lms/$lmsCode/modules/create',
+)({
+  component: RouteComponent,
+})
+
+function RouteComponent() {
+  const { authUser } = useAuth('protected')
+  
+  return (
+    <RoleComponentManager
+      currentRole={authUser.role}
+      roleRender={{
+        admin: <ModulesCreateAdminPage />,
+      }}
+    />
+  )
+}
