+++ frontend/src/routes/(protected)/lms/$lmsCode/modules/edit.tsx
@@ -1,0 +1,23 @@
+import RoleComponentManager from '@/components/role-component-manager'
+import { useAuth } from '@/features/auth/auth.hook'
+import ModulesCreateAdminPage from '@/pages/admin/lms/$lmsCode/modules/create'
+import { createFileRoute } from '@tanstack/react-router'
+
+export const Route = createFileRoute(
+  '/(protected)/lms/$lmsCode/modules/edit',
+)({
+  component: RouteComponent,
+})
+
+function RouteComponent() {
+  const { authUser } = useAuth('protected')
+
+  return (
+    <RoleComponentManager
+      currentRole={authUser.role}
+      roleRender={{
+        admin: <ModulesCreateAdminPage />,
+      }}
+    />
+  )
+}
