+++ frontend/src/routes/(protected)/lms/$lmsCode/overview/index.tsx
@@ -1,0 +1,12 @@
+import CourseOverview from '@/pages/shared/lms/$lmsCode/overview'
+import { createFileRoute } from '@tanstack/react-router'
+
+export const Route = createFileRoute(
+  '/(protected)/lms/$lmsCode/overview/',
+)({
+  component: RouteComponent,
+})
+
+function RouteComponent() {
+  return <CourseOverview />
+}
